Move `Merge` trait to a separate file
